requirements.txt
asgiref==3.4.1
Django==3.2.7
django-filter==21.1
djangorestframework==3.12.4
djangorestframework-simplejwt==5.0.0
Markdown==3.3.4
Pillow==8.3.2
psycopg2==2.9.1
PyJWT==2.2.0
python-dotenv==0.19.0
pytz==2021.1
sqlparse==0.4.2
